MFZE1
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@90647 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_hash.c
@@ -393,7 +393,7 @@ ZEND_API int zend_hash_index_update_or_next_insert(HashTable *ht, ulong h, void 
 			}
 			UPDATE_DATA(ht, p, pData, nDataSize);
 			HANDLE_UNBLOCK_INTERRUPTIONS();
-			if (h >= ht->nNextFreeElement) {
+			if ((long)h >= (long)ht->nNextFreeElement) {
 				ht->nNextFreeElement = h + 1;
 			}
 			if (pDest) {
@@ -421,7 +421,7 @@ ZEND_API int zend_hash_index_update_or_next_insert(HashTable *ht, ulong h, void 
 	CONNECT_TO_GLOBAL_DLLIST(p, ht);
 	HANDLE_UNBLOCK_INTERRUPTIONS();
 
-	if (h >= ht->nNextFreeElement) {
+	if ((long)h >= (long)ht->nNextFreeElement) {
 		ht->nNextFreeElement = h + 1;
 	}
 	ht->nNumOfElements++;
